requirements.txt
datasets
pillow
python-osc
unidecode
sentencepiece

